Aggregate API tests
